Add file t/extlib/.nosearch to keep t/extlib of the load-path
It is generally a bad idea to bundle libraries, including test
frameworks.  However since `el-expectations' is obsolete (its author
went on to write `ert-expectations') and is not available from Melpa
or another `package.el' archive, it does make sense here.

Adding a file named ".nosearch" keeps the containing directory of the
`load-path'.  And in this case it also prevents dependency extraction
tools (such as those used to maintain the Emacsmirror) from thinking
that `el-expectations' is part of `cake2'.